+++ code/mobile/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~4.0.1",
+    "@expo/ngrok": "^4.1.0",
     "@rn-primitives/alert-dialog": "^1.1.0",
     "@rn-primitives/label": "^1.1.0",
     "@rn-primitives/portal": "^1.1.0",
@@ -27,7 +28,7 @@
     "@trpc/server": "^11.1.4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "expo": "~52.0.38",
+    "expo": "~53.0.0",
     "expo-constants": "~17.0.8",
     "expo-linking": "~7.0.5",
     "expo-router": "~4.0.18",
